Refactor controllers to have only one controller file.
